Refactor to group together deep search components
